Scheduled Update
The clock adjust function is finally complete, now its possible to change date as well as time, the needed logic for adjusting days and months has been completed, now months with 31 days will have 31 days to choose, the same months with 30 days and february. Also, when a month is change, if the current selected day is above the new month's limit, its going to be automatically changed.
Started the process to change the menu for a new version, functionality has yet to be changed on most options and the first is incomplete.