Scheduled Update
Now every time a reading is taken it is saved to an address in the EEPROM. The history has been worked on, it is for now displaying the first three readings, using their date and time. The clean history option is now working and offering the % of work done so the user may follow it.