Scheduled Update
Worked a little more on how the history option works, especially attempting to make it smother to roll around the menu(not finished yet).
Now all the readings have a detailed page that can be assessed as well, and the history menu is also not wasting time by 'getting' the reading every single loop(have yet to finish the code around this).
Now when a new reading is taken, the program will look at the eeLimit variable to know if there is space for another reading before attempting to simply save it.