Scheduled Update
Some more work was done to the new menu function, now using a single switch case for naviagation, the up and down buttons were finally renamed from the original left and right, the global variables pwr and pct were moved inside the upperbar function and the scrollbar function was both updated for the new menu as well as removed from the upperbar. a new printOption was also created to print the options inside the history menu, making it a lot more compact.