Scheduled Update
Now it is possible to choose what frequency to use when taking readings, also, the frequency is also now possible to be seen on the details screen in the history. the frequency variable inside the reading struct was also changed to a long, this has added an aditional 2 bytes to the size of the struct, similar changes were made when saving a reading to the eeprom, the sincronization option is now working, however it only begins/ends the serial connection between arduino and bluetooth. there was a mistake made when choosing the positioning in the clockadjust function that was corrected, similarly, a piece of old code was removed to make the blinker function to work. the custom sweep function was adjusted to take a new frequency before performing the sweep, any needed change for this was also made.