Scheduled Update
Created the red_fang library, this is the library that interacts with the bluetooth, all but the REQ and GET functions are working for now.
the red_fang library is also being called directly on the arduino, receiving the red_beryl as a parameter because of how it must call some functions.
changed the update timer inside upperbar function to 1/4 second.