Scheduled Update
Corrected a strange bug where the pwr variable was sometimes receiving -1 or 0, also, adjusted the time between battery updates, as well as correcting the 99/100 flickering that sometimes happened.
Added the code for Req and Get in the red_fang library, both are working as intended. with this the red_fang library should be finished.
Completed the new reading option with the option to choose a frequency, now it will not use the default config anymore.