Scheduled Update
The GET command now has an alternative GTX(GeT CompleX) that returns the complete values of a reading.
This means, of course, that the GET command only returns the averaged values of a reading and is the one that is supposed to be used usually.
Corrected another memory leak bug as well as the display error linked to it.
When returning a reading, the individual variables are not wrapped around [S][E] anymore, now only the entire reading is wrapped, this saves some bytes during transmission.
Finally added the changed AD5933 library to the github folder as well.